+++ src/App.jsx
@@ -1,9 +1,9 @@
-import ModalAddWater from './components/Modals/ModalAddWater';
-
+import ModalAddWater from './components/Modals/ModalAdd/ModalAddWater';
+import ModalEditWater from './components/Modals/EditWater/EditWater';
 
 export const App = () => {
   return <div>
-
+<ModalEditWater/>
     <ModalAddWater/>
     React Node Project
   </div>;
