+++ src/components/Modals/EditWater/EditWater.module.css
@@ -20,11 +20,14 @@ button {
 }
 
 .modalwrapper {
-  position: relative;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
   border-radius: 15px;
   padding: 40px;
   width: 518px;
-  height: 555px;
+  height: 587px;
   box-shadow: 0 4px 50px 0 rgba(0, 0, 0, 0.1);
   background: #fff;
 }
@@ -80,6 +83,9 @@ button {
 }
 
 .button_water {
+  display: inline-block;
+  margin: 0;
+  vertical-align: top;
   width: 43px;
   height: 43px;
   border-radius: 24px;
@@ -98,6 +104,7 @@ button {
 
 .button_ml {
   font-family: "Poppins", sans-serif;
+  display: inline-block;
   margin-left: 5px;
   margin-right: 5px;
   font-weight: 700;
