+++ src/components/Modals/ModalAdd/ModalAddWater.module.css
@@ -1,0 +1,146 @@
+h1,
+h2,
+h3 {
+  margin: 0;
+  padding: 0;
+}
+
+button {
+  border: 2px solid #fff;
+  cursor: pointer;
+  background: #fff;
+}
+
+.close_button {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  width: 28px;
+  height: 28px;
+}
+
+.modalwrapper {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  border-radius: 15px;
+  padding: 40px;
+  width: 518px;
+  height: 587px;
+  box-shadow: 0 4px 50px 0 rgba(0, 0, 0, 0.1);
+  background: #fff;
+}
+
+.button_wrapper {
+  display: flex;
+  flex-direction: column;
+}
+
+.title {
+  font-family: "Poppins", sans - serif;
+  font-weight: 700;
+  font-size: 28px;
+  line-height: 1.14286;
+  letter-spacing: -0.01em;
+  color: #323f47;
+  padding-bottom: 24px;
+}
+
+.subtitle {
+  font-family: "Poppins", sans - serif;
+  padding-bottom: 20px;
+  font-weight: 700;
+  font-size: 18px;
+  line-height: 1.33333;
+  letter-spacing: -0.01em;
+  color: #323f47;
+}
+
+.amount_water {
+  font-family: "Poppins", sans - serif;
+  padding-bottom: 10px;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.5;
+  letter-spacing: -0.01em;
+  color: #2f2f2f;
+}
+
+.button_save {
+  border-radius: 30px;
+  padding: 18px 50px;
+  width: 141px;
+  height: 60px;
+  background: #9be1a0;
+  font-family: "Poppins", sans - serif;
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 1.5;
+  letter-spacing: -0.01em;
+  text-align: center;
+  color: #323f47;
+}
+
+.button_water {
+  display: inline-block;
+  margin: 0;
+  vertical-align: top;
+  width: 43px;
+  height: 43px;
+  border-radius: 24px;
+  border: 1.5px solid #323f47;
+  font-size: 24px;
+  background: #fff;
+  color: #323f47;
+  font-size: 24px;
+}
+
+.pl_min {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.button_ml {
+  font-family: "Poppins", sans-serif;
+  display: inline-block;
+  margin-left: 5px;
+  margin-right: 5px;
+  font-weight: 700;
+  font-size: 15px;
+  color: #fff;
+  border-radius: 30px;
+  width: 86px;
+  height: 43px;
+  background: #323f47;
+}
+
+.waterwrapper {
+  width: 188px;
+  padding-bottom: 16px;
+}
+
+input {
+  padding-right: 15px;
+  padding-left: 16px;
+  margin-bottom: 24px;
+  border: 1px solid rgba(47, 47, 47, 0.15);
+  border-radius: 15px;
+  height: 56px;
+  background: #fff;
+  font-family: "Poppins", sans - serif;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.5;
+  letter-spacing: -0.01em;
+  color: #2f2f2f;
+}
+
+.closebutton {
+  stroke-width: 3px;
+  stroke: #2f2f2f;
+  background: #fff;
+  border-color: #fff;
+  cursor: pointer;
+}
